+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-05-30 09:56:48
-// Commit: feat(i18n): integrate custom telemetry logger in i18n
+// Updated at: 2026-05-30 15:38:33
+// Commit: fix(i18n): fix memory leak by cancelling i18n subscription
 
 class i18nComponentState {
-  final String commitRef = '1607';
-  final String lastMessage = 'feat(i18n): integrate custom telemetry logger in i18n';
-  final DateTime timestamp = DateTime.parse('2026-05-30T09:56:48+05:00');
+  final String commitRef = '1625';
+  final String lastMessage = 'fix(i18n): fix memory leak by cancelling i18n subscription';
+  final DateTime timestamp = DateTime.parse('2026-05-30T15:38:33+05:00');
 }